+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/lib/console_link/console_link/logic/backfill.py
@@ -1,0 +1,32 @@
+import logging
+from typing import Dict, Optional
+
+from console_link.models.backfill_osi import OpenSearchIngestionBackfill
+from console_link.models.backfill_rfs import DockerRFSBackfill, ECSRFSBackfill
+from console_link.models.cluster import Cluster
+
+logger = logging.getLogger(__name__)
+
+
+class UnsupportedBackfillTypeError(Exception):
+    def __init__(self, supplied_backfill: str):
+        super().__init__("Unsupported backfill type", supplied_backfill)
+
+
+def get_backfill(config: Dict, source_cluster: Optional[Cluster], target_cluster: Optional[Cluster]):
+    if 'opensearch_ingestion' in config:
+        return OpenSearchIngestionBackfill(config=config,
+                                           source_cluster=source_cluster,
+                                           target_cluster=target_cluster)
+    elif 'reindex_from_source' in config:
+        if 'docker' in config['reindex_from_source']:
+            return DockerRFSBackfill(config=config,
+                                     target_cluster=target_cluster)
+        elif 'ecs' in config['reindex_from_source']:
+            return ECSRFSBackfill(config=config,
+                                  target_cluster=target_cluster)
+
+    logger.error(f"An unsupported metrics source type was provided: {config.keys()}")
+    if len(config.keys()) > 1:
+        raise UnsupportedBackfillTypeError(', '.join(config.keys()))
+    raise UnsupportedBackfillTypeError(next(iter(config.keys())))
