+++ TrafficCapture/dockerSolution/src/main/docker/migrationConsole/lib/console_link/console_link/models/backfill_rfs.py
@@ -1,0 +1,9 @@
+from backfill_base import Backfill
+
+
+class DockerRFSBackfill(Backfill):
+    pass
+
+
+class ECSRFSBackfill(Backfill):
+    pass
